Create remaining html sections. Add most of css
The only html I have left is to add a border to the info-images and the sign up button.
I do need to find a few random images, and may add an image in the lower section background.

Most of the styling is setup, though I know it could be optimized and setup in better ways.